Updated news table into functional component
