+++ src/Components/actTable/actList.css
@@ -1,6 +1,9 @@
 .btnUpdate{
     color:green
 }
+.dsa{
+    border: none;
+}
 
 
 @media screen and (max-width: 600px) {
