+++ src/UI/activitiesList/UIactList.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import ActList from '../../components/actTable/actList';
+
+function UIactList(
+){
+    
+    return(
+    <div >
+        <ActList />
+    </div>
+    
+    );
+}
+export default UIactList;
